requirements.txt
# === FRAMEWORK PRINCIPAL ===
Flask==2.3.3
Werkzeug==2.3.7

# === BASE DE DATOS ===
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4

# === SEGURIDAD Y AUTENTICACIÓN ===
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
argon2-cffi==21.3.0

# === VALIDACIÓN DE DATOS ===
marshmallow==3.19.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# === SERVIDOR DE PRODUCCIÓN ===
waitress==2.1.2

# === UTILIDADES ===
requests==2.31.0
email-validator==2.0.0
